catch all exceptions not just webexceptions, some fixes

+++ src/SwiftClient.Cli/Program.cs
@@ -13,6 +13,7 @@ namespace SwiftClient.Cli
     {
         SwiftClient client = null;
         SwiftCredentials credentials = new SwiftCredentials();
+        long bufferSize = 1000000;
 
         public async Task Main(string[] args)
         {
@@ -115,7 +116,23 @@ namespace SwiftClient.Cli
             var response = new SwiftResponse();
             var fileName = Path.GetFileNameWithoutExtension(options.File);
             string containerTemp = options.Container + "_tmp";
-            byte[] buffer = new byte[10000];
+            byte[] buffer = new byte[bufferSize];
+
+            response = client.PutContainer(containerTemp).Result;
+
+            if (!response.IsSuccess)
+            {
+                Console.WriteLine($"Put container error {response.Reason}");
+                return 500;
+            }
+
+            response = client.PutContainer(options.Container).Result;
+
+            if (!response.IsSuccess)
+            {
+                Console.WriteLine($"Put container error {response.Reason}");
+                return 500;
+            }
 
             using (var stream = File.OpenRead(options.File))
             {
@@ -140,17 +157,29 @@ namespace SwiftClient.Cli
                 // use manifest to merge chunks
                 response = client.PutManifest(containerTemp, fileName).Result;
 
+                if (!response.IsSuccess)
+                {
+                    Console.WriteLine($"Put manifest error {response.Reason}");
+                    return 500;
+                }
+
                 // copy chunks to new file and set some meta data info about the file (filename, contentype)
-                client.CopyObject(containerTemp, fileName, options.Container, options.Object, new Dictionary<string, string>
+                response = client.CopyObject(containerTemp, fileName, options.Container, options.Object, new Dictionary<string, string>
                 {
                     { string.Format(SwiftHeaderKeys.ObjectMetaFormat, "Filename"), fileName },
                     { string.Format(SwiftHeaderKeys.ObjectMetaFormat, "Contenttype"), Path.GetExtension(options.File) }
-                }).Wait();
+                }).Result;
+
+                if (!response.IsSuccess)
+                {
+                    Console.WriteLine($"Copy object error {response.Reason}");
+                    return 500;
+                }
 
                 // cleanup temp chunks
-                var deleteTasks = new List<Task>();
+                var deleteTasks = new List<Task<SwiftResponse>>();
 
-                for (var i = 0; i <= chunks; i++)
+                for (var i = 0; i < chunks; i++)
                 {
                     deleteTasks.Add(client.DeleteObjectChunk(containerTemp, fileName, i));
                 }
@@ -159,7 +188,27 @@ namespace SwiftClient.Cli
                 deleteTasks.Add(client.DeleteObject(containerTemp, fileName));
 
                 // cleanup temp container
-                Task.WhenAll(deleteTasks).Wait();
+                Task.WhenAll(deleteTasks).ContinueWith((rsp) =>
+                {
+                    response = rsp.Result.FirstOrDefault(x => !x.IsSuccess);
+
+                    if (response != null)
+                    {
+                        Console.WriteLine($"Cleanup temp chunks error {response.Reason}");
+                    }
+
+                    if (rsp.Result.All(x => x.IsSuccess))
+                    {
+                        response = client.DeleteContainer(containerTemp).Result;
+                    }
+
+                }).Wait();
+
+                if (!response.IsSuccess)
+                {
+                    Console.WriteLine($"Cleanup error {response.Reason}");
+                    return 500;
+                }
 
                 Console.WriteLine($"Upload done");
             }
